+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: tortiz-r <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/16 13:13:20 by tortiz-r          #+#    #+#             */
-/*   Updated: 2024/02/16 18:59:38 by tortiz-r         ###   ########.fr       */
+/*   Updated: 2024/02/24 19:04:03 by tortiz-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_memset(void *b, int c, size_t len)
 	int	i;
 
 	i = 0;
-	while (i < len)
+	while ((unsigned long) i < len)
 	{
 		((char *)b)[i] = (char) c;
 		i++;
